Update UI caption and MaxCheckpoints calculation

+++ Code/main.lua
@@ -137,7 +137,7 @@ function InitTrackInfo()
   S1_raw=0
   S2_raw=0
   S3_raw=0
-  MaxCheckpoints = readFloat('GTA5.exe+201B9BC') - 1  --old CBF40 new D 16C0 del prev 97C60 928
+  MaxCheckpoints = readInteger('GTA5.exe+201B9BC') - 1  --old CBF40 new D 16C0 del prev 97C60 928
   local Track_Name = readString('adr + 6051E150')
   UI.Caption = Track_Name.." Delta App Online 1.2.7"
   CurLapLastCheckpointTime = 0
